Welcome Page WebUI: Move brave-ui paths to brave-core paths
