+++ components/brave_welcome_ui/components/images/index.ts
