+++ components/brave_welcome_ui/components/index.ts
