+++ components/brave_welcome_ui/components/select/index.ts
@@ -2,8 +2,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import styled, { css } from '../../../theme'
-import palette from '../../../theme/colors'
+import styled, { css } from 'brave-ui/theme'
+import palette from 'brave-ui/theme/colors'
 import caratUrl from './carat.svg'
 
 export interface SelectBoxProps {
