+++ components/brave_welcome_ui/components/wrapper/index.ts
