+++ components/brave_welcome_ui/containers/screens/themeBox.tsx
@@ -5,10 +5,10 @@
 import * as React from 'react'
 
 // Feature-specific components
-import { Content, Title, Paragraph, PrimaryButton } from 'brave-ui/features/welcome'
+import { Content, Title, Paragraph, PrimaryButton } from '../../components'
 
 // Images
-import { WelcomeThemeImage } from 'brave-ui/features/welcome/images'
+import { WelcomeThemeImage } from '../../components/images'
 
 // Utils
 import { getLocale } from '../../../common/locale'
