+++ components/brave_welcome_ui/stories/page/fakeLocale.ts
